requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
absl-py==1.4.0
adapter-transformers==3.2.1
cachetools==5.3.0
charset-normalizer==2.1.1
colorama==0.4.6
filelock==3.9.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
Pillow==9.3.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.3
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm==4.65.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.3